+++ lib/src/play_store_search_api.dart
@@ -7,6 +7,8 @@ import 'package:html/parser.dart' show parse;
 import 'package:http/http.dart' as http;
 import 'package:version/version.dart';
 
+// TODO: refactor the two store API classes to use more shared code.
+
 class PlayStoreSearchAPI {
   /// Play Store Search Api URL
   final String playStorePrefixURL = 'play.google.com';
@@ -58,6 +60,43 @@ class PlayStoreSearchAPI {
 }
 
 class PlayStoreResults {
+  /// Return field description from Play Store results.
+  static String? description(Document response) {
+    try {
+      final sectionElements = response.getElementsByClassName('W4P4ne');
+      final descriptionElement = sectionElements[0];
+      final description = descriptionElement
+          .querySelector('.PHBdkd')
+          ?.querySelector('.DWPxHb')
+          ?.text;
+      return description;
+    } catch (e) {
+      print('upgrader: PlayStoreResults.description exception: $e');
+    }
+    return null;
+  }
+
+  /// Return the minimum app version taken from the tag in the description field
+  /// from the store response. The format is: [:mav: 1.2.3].
+  /// Returns version, such as 1.2.3, or null.
+  static Version? minAppVersion(Document response, {String tagName = 'mav'}) {
+    var version;
+    try {
+      final description = PlayStoreResults.description(response);
+      if (description != null) {
+        final regExpSource = r'\[\:mav\:[\s]*(?<version>[^\s]+)[\s]*\]';
+        final regExp = RegExp(regExpSource, caseSensitive: false);
+        final match = regExp.firstMatch(description);
+        final mav = match != null ? match.namedGroup('version') : null;
+        // Verify version string using class Version
+        version = mav != null ? Version.parse(mav) : null;
+      }
+    } on Exception catch (e) {
+      print('upgrader.PlayStoreResults.minAppVersion : $e');
+    }
+    return version;
+  }
+
   /// Returns field releaseNotes from Play Store results. When there are no
   /// release notes, the main app description is used.
   static String? releaseNotes(Document response) {